Fixed formatting, added new tests
