+++ rid_operations/rid_utils.py
@@ -1,13 +1,9 @@
 import enum
-import json
 from dataclasses import asdict, dataclass, field
-from typing import List, Literal, NamedTuple, Optional, Tuple, Union
+from typing import List, Literal, NamedTuple, Optional, Union
 
-import arrow
-from arrow.parser import ParserError
 from implicitdict import StringBasedDateTime
 
-from auth_helper.common import get_redis
 from scd_operations.scd_data_definitions import Volume4D
 
 from .data_definitions import UASID, UAClassificationEU
@@ -117,8 +113,8 @@ class ISACreationRequest:
 class IdentificationServiceArea:
     uss_base_url: str
     owner: str
-    time_start: StringBasedDateTime
-    time_end: StringBasedDateTime
+    time_start: RIDTime
+    time_end: RIDTime
     version: str
     id: str
 
@@ -158,6 +154,12 @@ class RIDCapabilitiesResponse:
 
 
 @dataclass
+class RIDHeight:
+    distance: float
+    reference: str
+
+
+@dataclass
 class RIDAircraftPosition:
     lat: float
     lng: float
@@ -166,12 +168,8 @@ class RIDAircraftPosition:
     accuracy_v: str
     extrapolated: Optional[bool]
     pressure_altitude: Optional[float]
+    height: Optional[RIDHeight]
 
-
-@dataclass
-class RIDHeight:
-    distance: float
-    reference: str
 
 @dataclass
 class AuthData:
@@ -303,7 +301,7 @@ class TelemetryFlightDetails:
 
 @dataclass
 class RIDFlightResponse:
-    timestamp: str
+    timestamp: RIDTime
     flights: List[TelemetryFlightDetails]
 
 
@@ -311,134 +309,3 @@ class RIDFlightResponse:
 class SingleObservationMetadata:
     details_response: RIDTestDetailsResponse
     telemetry: RIDAircraftState
-
-
-def process_requested_flight(
-    requested_flight: dict,
-    flight_injection_sorted_set:str
-) -> tuple[RIDTestInjection, List[LatLngPoint], List[float]]:
-    r = get_redis()
-    all_telemetry = []
-    all_flight_details = []
-    all_positions: List[LatLngPoint] = []
-    all_altitudes = []
-    provided_telemetries = requested_flight["telemetry"]
-    provided_flight_details = requested_flight["details_responses"]
-
-    for provided_flight_detail in provided_flight_details:
-        fd = provided_flight_detail["details"]
-        requested_flight_detail_id = fd["id"]
-
-        op_location = LatLngPoint(lat=fd["operator_location"]["lat"], lng=fd["operator_location"]["lng"])
-        if "auth_data" in fd.keys():
-            auth_data = RIDAuthData(format=fd["auth_data"]["format"], data=fd["auth_data"]["data"])
-        else:
-            auth_data = RIDAuthData(format="0", data="")
-        serial_number = fd["serial_number"] if "serial_number" in fd else "MFR1C123456789ABC"
-        if "uas_id" in fd.keys():
-            uas_id = UASID(
-                specific_session_id=fd["uas_id"]["specific_session_id"],
-                serial_number=fd["uas_id"]["serial_number"],
-                registration_id=fd["uas_id"]["registration_id"],
-                utm_id=fd["uas_id"]["utm_id"],
-            )
-        else:
-            uas_id = UASID(
-                specific_session_id="02-a1b2c3d4e5f60708",
-                serial_number=serial_number,
-                utm_id="ae1fa066-6d68-4018-8274-af867966978e",
-                registration_id="MFR1C123456789ABC",
-            )
-        if "eu_classification" in fd.keys():
-            eu_classification = UAClassificationEU(
-                category=fd["eu_classification"]["category"],
-                class_=fd["eu_classification"]["class"],
-            )
-        else:
-            eu_classification = UAClassificationEU(category="EUCategoryUndefined", class_="EUClassUndefined")
-
-        flight_detail = RIDOperatorDetails(
-            id=requested_flight_detail_id,
-            operation_description=fd["operation_description"],
-            serial_number=serial_number,
-            registration_number=fd["registration_number"],
-            operator_location=op_location,
-            aircraft_type="NotDeclared",
-            operator_id=fd["operator_id"],
-            auth_data=auth_data,
-            uas_id=uas_id,
-            eu_classification=eu_classification,
-        )
-        pfd = RIDTestDetailsResponse(
-            effective_after=provided_flight_detail["effective_after"],
-            details=flight_detail,
-        )
-        all_flight_details.append(pfd)
-
-        flight_details_storage = "flight_details:" + requested_flight_detail_id
-
-        r.set(flight_details_storage, json.dumps(asdict(flight_detail)))
-        # expire in 5 mins
-        r.expire(flight_details_storage, time=3000)
-
-    # Iterate over telemetry details provided
-    for telemetry_id, provided_telemetry in enumerate(provided_telemetries):
-        pos = provided_telemetry["position"]
-        # In provided telemetry position and pressure altitude and extrapolated values are optional use if provided else generate them.
-        pressure_altitude = pos["pressure_altitude"] if "pressure_altitude" in pos else 0.0
-        extrapolated = pos["extrapolated"] if "extrapolated" in pos else False
-
-        llp = LatLngPoint(lat=pos["lat"], lng=pos["lng"])
-        all_positions.append(llp)
-        all_altitudes.append(pos["alt"])
-        position = RIDAircraftPosition(
-            lat=pos["lat"],
-            lng=pos["lng"],
-            alt=pos["alt"],
-            accuracy_h=pos["accuracy_h"],
-            accuracy_v=pos["accuracy_v"],
-            extrapolated=extrapolated,
-            pressure_altitude=pressure_altitude,
-        )
-
-        if "height" in provided_telemetry.keys():
-            height = RIDHeight(
-                distance=provided_telemetry["height"]["distance"],
-                reference=provided_telemetry["height"]["reference"],
-            )
-        else:
-            height = None
-
-        try:
-            formatted_timestamp = arrow.get(provided_telemetry["timestamp"])
-        except ParserError:
-            logger.info("Error in parsing telemetry timestamp")
-        else:
-            teletemetry_observation = RIDAircraftState(
-                timestamp=RIDTime(value=provided_telemetry["timestamp"], format="RFC3339"),
-                timestamp_accuracy=provided_telemetry["timestamp_accuracy"],
-                operational_status=provided_telemetry["operational_status"],
-                position=position,
-                track=provided_telemetry["track"],
-                speed=provided_telemetry["speed"],
-                speed_accuracy=provided_telemetry["speed_accuracy"],
-                vertical_speed=provided_telemetry["vertical_speed"],
-                height=height,
-            )
-            closest_details_response = min(
-                all_flight_details,
-                key=lambda d: abs(arrow.get(d.effective_after) - formatted_timestamp),
-            )
-            flight_state_storage = RIDTestDataStorage(flight_state=teletemetry_observation, details_response=closest_details_response)
-            zadd_struct = {json.dumps(asdict(flight_state_storage)): formatted_timestamp.int_timestamp}
-            # Add these as a sorted set in Redis
-            r.zadd(flight_injection_sorted_set, zadd_struct)
-            all_telemetry.append(teletemetry_observation)
-
-    _requested_flight = RIDTestInjection(
-        injection_id=requested_flight["injection_id"],
-        telemetry=all_telemetry,
-        details_responses=all_flight_details,
-    )
-
-    return _requested_flight, all_positions, all_altitudes
